requirements.txt
os
requests
google-cloud-storage